Revert "Ported in fix from master for default delay value type casting"
This reverts commit c53490bed1884071da367b3dc183c7739754b891.

So we can be sure we're testing against the original encoder this is being temporarily reverted.

+++ Encoder/src/main/java/com/hak5/usbrubberducky/Encoder.java
@@ -155,7 +155,7 @@ public class Encoder {
 
                 if (instruction[0].equals("DEFAULT_DELAY")
                         || instruction[0].equals("DEFAULTDELAY")) {
-                    defaultDelay = Integer.parseInt(instruction[1]
+                    defaultDelay = (byte) Integer.parseInt(instruction[1]
                             .trim());
                 } else if (instruction[0].equals("DELAY")) {
                     int delay = Integer.parseInt(instruction[1].trim());
